TransitionableTransform: Fix halt()
Each final state needs to be set to the current state after the transition has been halted. Otherwise the _final* states do not represent the actual final states, but the final states of the previous (unfinished) transformation. This results into unexpected behavior when doing a single transformation (e.g. rotation) on a halted TransitionableTransform that didn't update its final states, since the _buildFinal uses the _final* of each Transformation (which hasn't been updated).