[Fix] Wrap Fast Click in Closure
Currently fast click calls return if ```!window.CustomEvent```.  This caused no problems when we were only releasing using AMD, as this return was wrapped in closure, but it unfortunately breaks everything when the framework is run in Common.

This fix wraps all the code inside of fast click in a self executing function, making sure that the return doesn't make everything throw up.