Relative path all the things
Moving forward we are planning to have famous all in a single src directory.  This is moving even further from the initial design of having git submodules.  As such maintaining absolute path's to modules is not only not necessary, but is actually incorrect and breaking certain tools.

This pull request goes through and replaces all references to famous with the appropriate relative path.